feat: revamp whole design
The contract now doesn't create any streams at all. It allows a seller to create an agreement. A buyer can accept the agreement by first creating a stream and then calling the acceptAgreement function. That will verify that the correct stream was created and accept the agreement. Besides that, there will be the logic of transferring the NFT and the deposit as needed.

There were two ideas:
1. contract creates stream for the buyer -> technically not feasible
2. buyer creates stream to contract and contract creates stream to
seller -> all sorts of issues with gas costs for the contract etc. We
would have to implement a fee for the contract to have the funds to
create streams all the time.

New solution: let the buyer handle the stream stuff outside our
contract. Contract only handles the locking of a deposit and the NFT and
verifies that the stream was executed successfully. Nothign else.

Makes it more simple.